Transferred Search bar methods from Main Activity to Search
