+++ .idea/modules.xml
@@ -3,6 +3,7 @@
   <component name="ProjectModuleManager">
     <modules>
       <module fileurl="file://$PROJECT_DIR$/ParkIt.iml" filepath="$PROJECT_DIR$/ParkIt.iml" />
+      <module fileurl="file://$PROJECT_DIR$/ParkIt.iml" filepath="$PROJECT_DIR$/ParkIt.iml" />
       <module fileurl="file://$PROJECT_DIR$/ParkIt-2.iml" filepath="$PROJECT_DIR$/ParkIt-2.iml" />
       <module fileurl="file://$PROJECT_DIR$/app/app.iml" filepath="$PROJECT_DIR$/app/app.iml" />
     </modules>
