+++ app/src/main/AndroidManifest.xml
@@ -26,7 +26,7 @@
 
         <meta-data
             android:name="com.google.android.geo.API_KEY"
-            android:value="AIzaSyAWmkhxzlMzF9IFJpCh-bDs_fKh3YATSgA"/>
+            android:value="GOOGLE_MAPS_API_KEY"/>
 
         <activity
             android:name="com.csc413.group9.parkIt.MainActivity"
