--- app/src/main/java/com/csc413/group9/parkIt/Features/Search.java
@@ -1,7 +1,0 @@
-package com.csc413.group9.parkIt.Features;
-
-/**
- * Created by kshen on 4/24/15.
- */
-public class Search {
-}
